requirements.txt
werkzeug
Flask==1.0.3
python-dotenv==0.10.3
psycopg2==2.8.2
pytest==4.6.3
requests==2.22.0
